WIP on supporting ST4 side-by-side opening
This allows to open a the highlighted selection entry to be opened next
to the current. This tries to behave in the same way as the default
file opener of ST4.

Please note that this will only work with current ST4 builds and there
is no fallback/version check yet.